+++ demo-plugin/source/editor/jumpDemo_LookAndFeel.h
@@ -14,7 +14,7 @@ namespace jumpDemo
             void drawSeparator(juce::Graphics& g, const jump::Canvas& canvas) const noexcept override;
 
             jump::TrackInfoList getTemplateColumns() const noexcept override;
-            jump::Size<int> getDemoSelectorSize() const noexcept override;
+            jump::Size<float> getDemoSelectorSize() const noexcept override;
             jump::Margin<float> getDemoSelectorMargin() const noexcept override;
         };
     }
